requirements.txt
pandas==2.3.1
Pillow==11.3.0
python-dotenv==1.1.1
streamlit==1.45.1
streamlit_chat==0.1.1
streamlit_feedback==0.1.4
openai==1.98.0
openpyxl==3.1.5
db-dtypes==1.4.3
google-auth==2.40.3
google-cloud-bigquery==3.36.0
PyMySQL==1.1.2
python-docx==0.8.11
flask==3.1.2
flask-cors==6.0.1
SQLAlchemy==2.0.21
httpx==0.28.1
gunicorn==23.0.0